+++ mayan/apps/logging/tests/mixins.py
@@ -2,10 +2,27 @@ from django.apps import apps
 from django.contrib.auth import get_user_model
 from django.contrib.contenttypes.models import ContentType
 
-from mayan.apps.tests.tests.base import GenericViewTestCase
-
 from ..classes import ErrorLog
-from ..permissions import permission_error_log_view
+
+from .literals import TEST_ERROR_LOG_ENTRY_RESULT
+
+
+class LoggingTextMixin:
+    def _create_test_object(self):
+        self._create_test_user()
+        self.test_model = get_user_model()
+        self.test_object = self.test_user
+
+    def _create_error_log_entry(self):
+        app_config = apps.get_app_config(app_label='logging')
+        self.error_log = ErrorLog(app_config=app_config)
+        self.error_log.register_model(
+            model=self.test_model, register_permission=True
+        )
+
+        self.error_log_entry = self.test_object.error_log.create(
+            text=TEST_ERROR_LOG_ENTRY_RESULT
+        )
 
 
 class LoggingViewTestMixin: